Try to link PRs to issues mentioned in the PR title Fixes #3739
