+++ src/github/utils.ts
@@ -38,6 +38,9 @@ import {
 import { IssueModel } from './issueModel';
 import { GHPRComment, GHPRCommentThread } from './prComment';
 
+export const ISSUE_EXPRESSION = /(([A-Za-z0-9_.\-]+)\/([A-Za-z0-9_.\-]+))?(#|GH-)([1-9][0-9]*)($|\b)/;
+export const ISSUE_OR_URL_EXPRESSION = /(https?:\/\/github\.com\/(([^\s]+)\/([^\s]+))\/([^\s]+\/)?(issues|pull)\/([0-9]+)(#issuecomment\-([0-9]+))?)|(([A-Za-z0-9_.\-]+)\/([A-Za-z0-9_.\-]+))?(#|GH-)([1-9][0-9]*)($|\b)/;
+
 export interface CommentReactionHandler {
 	toggleReaction(comment: vscode.Comment, reaction: vscode.CommentReaction): Promise<void>;
 }
@@ -48,6 +51,27 @@ export type ParsedIssue = {
 	issueNumber: number;
 	commentNumber?: number;
 };
+
+export function parseIssueExpressionOutput(output: RegExpMatchArray | null): ParsedIssue | undefined {
+	if (!output) {
+		return undefined;
+	}
+	const issue: ParsedIssue = { owner: undefined, name: undefined, issueNumber: 0 };
+	if (output.length === 7) {
+		issue.owner = output[2];
+		issue.name = output[3];
+		issue.issueNumber = parseInt(output[5]);
+		return issue;
+	} else if (output.length === 16) {
+		issue.owner = output[3] || output[11];
+		issue.name = output[4] || output[12];
+		issue.issueNumber = parseInt(output[7] || output[14]);
+		issue.commentNumber = output[9] !== undefined ? parseInt(output[9]) : undefined;
+		return issue;
+	} else {
+		return undefined;
+	}
+}
 
 export function threadRange(startLine: number, endLine: number, endCharacter?: number): vscode.Range {
 	if ((startLine !== endLine) && (endCharacter === undefined)) {
