+++ src/issues/issueFeatureRegistrar.ts
@@ -10,7 +10,7 @@ import { OctokitCommon } from '../github/common';
 import { FolderRepositoryManager, PullRequestDefaults } from '../github/folderRepositoryManager';
 import { IssueModel } from '../github/issueModel';
 import { RepositoriesManager } from '../github/repositoriesManager';
-import { getRepositoryForFile } from '../github/utils';
+import { getRepositoryForFile, ISSUE_OR_URL_EXPRESSION, parseIssueExpressionOutput } from '../github/utils';
 import { ReviewManager } from '../view/reviewManager';
 import { CurrentIssue } from './currentIssue';
 import { IssueCompletionProvider } from './issueCompletionProvider';
@@ -35,10 +35,8 @@ import {
 	createGitHubLink,
 	createGithubPermalink,
 	getIssue,
-	ISSUE_OR_URL_EXPRESSION,
 	ISSUES_CONFIGURATION,
 	NewIssue,
-	parseIssueExpressionOutput,
 	PermalinkInfo,
 	pushAndCreatePR,
 	QUERIES_CONFIGURATION,
